refactor: 外移 shared_sessions/skills_hub 内联测试
- shared_sessions.rs 2816→2061（-755）；skills_hub.rs 3574→3062（-512）
- 纯搬运零行为变更；skills_hub env 隔离设置顺序原样保持
- baseline 重生成 + 新文件登记豁免
验证：cargo test shared_sessions 29 passed、skills_hub 16 passed；cargo check 通过；gate 绿
