+++ docs/architecture/large-file-baseline.json
@@ -1,5 +1,5 @@
 {
-  "generatedAt": "2026-08-30T14:57:45.869Z",
+  "generatedAt": "2026-08-30T14:59:25.909Z",
   "scope": "fail",
   "policyVersion": "2026-08-30.policy-v5",
   "entries": [
@@ -151,16 +151,6 @@
       "priority": "P2",
       "warnThreshold": 2600,
       "failThreshold": 3000,
-      "thresholdSource": "policy"
-    },
-    {
-      "path": "src-tauri/src/skills_hub.rs",
-      "lines": 3574,
-      "type": "rust",
-      "policyId": "bridge-runtime-critical",
-      "priority": "P0",
-      "warnThreshold": 2200,
-      "failThreshold": 2600,
       "thresholdSource": "policy"
     },
     {
@@ -404,6 +394,16 @@
       "thresholdSource": "policy"
     },
     {
+      "path": "src-tauri/src/skills_hub.rs",
+      "lines": 3062,
+      "type": "rust",
+      "policyId": "bridge-runtime-critical",
+      "priority": "P0",
+      "warnThreshold": 2200,
+      "failThreshold": 2600,
+      "thresholdSource": "policy"
+    },
+    {
       "path": "src/styles/messages.part1.css",
       "lines": 3035,
       "type": "css",
@@ -466,16 +466,6 @@
     {
       "path": "src-tauri/src/types.rs",
       "lines": 2828,
-      "type": "rust",
-      "policyId": "bridge-runtime-critical",
-      "priority": "P0",
-      "warnThreshold": 2200,
-      "failThreshold": 2600,
-      "thresholdSource": "policy"
-    },
-    {
-      "path": "src-tauri/src/shared_sessions.rs",
-      "lines": 2816,
       "type": "rust",
       "policyId": "bridge-runtime-critical",
       "priority": "P0",
